Remove all unnecessary functions - Add fesult props to song detail
