--- src/logic/mstoseg.js
@@ -1,7 +1,0 @@
-const miliToFormat = millis => {
-  const minutes = Math.floor(millis / 60000);
-  const seconds = ((millis % 60000) / 1000).toFixed(0);
-  return `${minutes}:${seconds < 10 ? `0${seconds}` : seconds}`;
-};
-
-export default miliToFormat;
