--- src/logic/sanitizeTrackName.js
@@ -1,5 +1,0 @@
-const sanitizeName = (trackName = 'van halen') => trackName
-  .replace(/-.*/, '')
-  .trim();
-
-export default sanitizeName;
