--- src/logic/searchVanHalenBand.js
@@ -1,3 +1,0 @@
-const searchVanHalenBand = array => array.filter(data => data.result.primary_artist.name === 'Van Halen');
-
-export default searchVanHalenBand;
